refactor: increasing hyperparameter epochs to 2500
As can be seen from the validation curve, both training and validation
curves are still decreasing, therefore we would need a few more epochs.
The number of epochs should be chosen such that the training error is
decreasing or has flattened out and the validation error has flattened
out and is not increasing. I think 2500 epochs should work fine here